requirements.txt
apache-airflow==2.8.1
psycopg2-binary
requests
python-dotenv
